+++ clients/java/src/main/java/io/lakefs/clients/sdk/model/Commit.java
@@ -90,7 +90,7 @@ public class Commit {
 
   public static final String SERIALIZED_NAME_VERSION = "version";
   @SerializedName(SERIALIZED_NAME_VERSION)
-  private Long version;
+  private Integer version;
 
   public Commit() {
   }
@@ -279,7 +279,7 @@ public class Commit {
   }
 
 
-  public Commit version(Long version) {
+  public Commit version(Integer version) {
     
     this.version = version;
     return this;
@@ -287,15 +287,17 @@ public class Commit {
 
    /**
    * Get version
+   * minimum: 0
+   * maximum: 1
    * @return version
   **/
   @javax.annotation.Nonnull
-  public Long getVersion() {
+  public Integer getVersion() {
     return version;
   }
 
 
-  public void setVersion(Long version) {
+  public void setVersion(Integer version) {
     this.version = version;
   }
